компил

+++ src/server/newscripts/Pandaria/SiegeofOrgrimmar/instance_siege_of_orgrimmar.cpp
@@ -328,7 +328,7 @@ public:
                 Map::PlayerList const& players = instance->GetPlayers();
                 if (!players.isEmpty())
                     if (Player* player = players.begin()->getSource())
-                        static_cast<uint32>(_teamInInstance) = player->GetTeam();
+                        _teamInInstance = player->GetTeam();
             }
 
             switch (creature->GetEntry())
@@ -1273,7 +1273,7 @@ public:
                         Map::PlayerList const& players = instance->GetPlayers();
                         if (!players.isEmpty())
                             if (Player* player = players.begin()->getSource())
-                                static_cast<uint32>(_teamInInstance) = player->GetTeam();
+                                _teamInInstance = player->GetTeam();
                     }
                     return _teamInInstance;
